PQNs are actually integers, cache the effective PQNs in the loop - simplified slaterPoint.
git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@1803 af78ad9f-9711-0410-ba57-a8eff4914ec4

+++ libavogadro/src/extensions/mopacaux.h
@@ -67,7 +67,7 @@ namespace Avogadro
     std::vector<int> m_atomIndex;
     std::vector<int> m_atomSym;
     std::vector<double> m_zeta;
-    std::vector<double> m_pqn;
+    std::vector<int> m_pqn;
     std::vector<Eigen::Vector3d> m_atomPos;
 
     Eigen::MatrixXd m_overlap;     /// Overlap matrix
